# apidoc/openapi.yaml
openapi: 3.0.1
paths:
  /application.wadl:
    get:
      operationId: getWadl
      responses:
        default:
          description: default response
          content:
            application/vnd.sun.wadl+xml: {}
            application/xml: {}
  /application.wadl/{path}:
    get:
      operationId: getExternalGrammar
      parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/xml: {}
  /info:
    get:
      operationId: getInfoForAll
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: string
  /info/all:
    get:
      operationId: allUsers
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: string
  /info/admin:
    get:
      operationId: getFromAdmin
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: string
  /info/user:
    get:
      operationId: getFromUser
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: string
  /Example/Authentication:
    get:
      tags:
      - user
      - login
      - GET
      summary: A method that checks if signed in as Admin.
      description: 'Returns a message if logged in as user '
      operationId: getFromAdmin_1
      responses:
        default:
          description: Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        403:
          description: Not logged in as admin
        200:
          description: succes
  /Example/postExample:
    post:
      tags:
      - POST
      summary: A method that is a simple post example.
      description: 'Returns a message if logged in as user '
      operationId: vote
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                param1:
                  type: string
                param2:
                  type: string
      responses:
        default:
          description: Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        403:
          description: wrong parameter names
        200:
          description: succes
  /Example/user:
    get:
      tags:
      - Authentication
      summary: A method that checks if signed in as user.
      description: 'Returns a message if logged in as user '
      operationId: getFromUser_1
      responses:
        default:
          description: Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        403:
          description: Not logged in as user
        200:
          description: succes
  /Example/{id}:
    get:
      tags:
      - POST
      summary: A method that is a simple get example with a parameter.
      description: 'Returns a message if logged in as user '
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        default:
          description: Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        404:
          description: Not a integer parameter
        200:
          description: succes
  /login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              type: string
      responses:
        default:
          description: default response
          content:
            application/json: {}
components:
  schemas:
    Message:
      type: object
      properties:
        msg:
          type: string
          description: Message from the server
      description: Message from server.